backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.39.0
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.23
aiohttp==3.9.1
python-multipart==0.0.6
openpyxl==3.1.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-telegram-bot==20.7
pytest==7.4.3
pytest-asyncio==0.21.1 